+++ src/main/java/beachcombing/backend/domain/beach/controller/dto/BeachFineMyMarkerResponse.java
@@ -1,17 +1,15 @@
-package beachcombing.backend.domain.beach.dto;
+package beachcombing.backend.domain.beach.controller.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-public class BeachFindMarkerResponse {
+public class BeachFineMyMarkerResponse {
     public Long id;
     public String name;
     public String lat;
     public String lng;
-    public String memberImage; // 최근 청소기록 유저 프로필이미지
 }
