+++ src/main/java/beachcombing/backend/domain/beach/service/helper/RayCastingHelper.java
@@ -1,0 +1,7 @@
+package beachcombing.backend.domain.beach.service.helper;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class RayCastingHelper {
+}
