+++ src/main/java/beachcombing/backend/domain/record/controller/RecordController.java
@@ -1,6 +1,9 @@
 package beachcombing.backend.domain.record.controller;
 
-import beachcombing.backend.domain.record.dto.RecordByBeachFindAllResponse;
+import beachcombing.backend.domain.record.controller.dto.RecordByBeachFindAllResponse;
+import beachcombing.backend.domain.record.controller.dto.RecordFindAllResponse;
+import beachcombing.backend.domain.record.controller.dto.RecordSaveRequest;
+import beachcombing.backend.domain.record.controller.dto.RecordSaveResponse;
 import beachcombing.backend.domain.record.dto.*;
 import beachcombing.backend.domain.record.service.RecordService;
 import beachcombing.backend.global.security.auth.PrincipalDetails;
