+++ src/main/java/beachcombing/backend/domain/refresh_token/domain/repository/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package beachcombing.backend.domain.refresh_token.repository;
+package beachcombing.backend.domain.refresh_token.domain.repository;
 
 import beachcombing.backend.domain.refresh_token.domain.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
